Fixed a small bug in attachments block, caused by the new addition of filetypes; added forum select box functions to includes/acp/acp_portal.php, which will be added to the ACP front-end in the next few days; Fixed small bugs for better compatibility with MSSQL;

+++ root/install/install_functions.php
@@ -154,7 +154,7 @@ function b3p_drop_table($table)
 
 	$table_name = substr($table . '#', 6, -1);
 
-	if ($db->sql_layer != 'mssql')
+	if ($db->sql_layer != 'mssql' && $db->sql_layer != 'mssql_odbc')
 	{
 		$sql = 'DROP TABLE IF EXISTS ' . $table_prefix . $table_name;
 		$result = $db->sql_query($sql);
